Setup use-package :defer correctly
If I just use :defer, then it waits to be loaded. This is fine for
libraries that re guaranteed to be loaded by another library. But, for
standanlone packages (without bindings or commands), I need to supply
a delay parameter

Most of these are for libraries that plug into existing systems (such
as the org exporters)